getTile() with quad tree caching.
This CR fixes the issue with the fetching quad tree index
with depth 0.

Fetching quad tree index for parent Tile of the requested Tile
with delta 4 is useful for caching all sub quads
and parent quads metadata for the later calls.

Returns the blob of the requested Tile or the blob
of the closest parent Tile.

Also changed getData method in Versioned and Volatile layer clients.
The method caches quad tree index with delta 4
and returns data of the requested tile or rejects if now found.

Added getAggregatedData method to the Versioned and
Volatile layer clients.
The method re-uses getTile function to return data
of the requested tile or its closest ancestor.

Added the integration test for testing getting tile and caching.

Resolves: OLPEDGE-1953, OLPEDGE-1893

Signed-off-by: Oleksii Zubko <ext-oleksii.zubko@here.com>